refactored and decoupled tuple_embedding.py
