+++ domain/estimators/__init__.py
@@ -2,5 +2,4 @@ from .logistic import Logistic
 from .naive_bayes import NaiveBayes
 from .tuple_embedding import TupleEmbedding
 
-
 __all__ = ['Logistic', 'NaiveBayes', 'TupleEmbedding']
