+++ domain/estimators/embedding/__init__.py
@@ -1,0 +1,6 @@
+from .dataset import LookupDataset
+from .model import EmbeddingModel
+from .predictor import Predictor
+from .trainer import Trainer
+
+__all__ = ['LookupDataset', 'EmbeddingModel', 'Predictor', 'Trainer']
